requirements.txt
fastapi==0.112.1
uvicorn==0.30.1
pydantic==2.8.2
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.1